move test to critical
